Add the proper commit() call on target index after adding contents of the secondary index.
I managed to reproduce the problem, and it was caused 
by a race in between the times the primary index is cleared 
(marks it dirty) and the adding of contents is finished.
If queries are done in this window, they will see the
empty index, because of the dirty-marking during clearing. But
no dirty-marking was done after adding the contents .. 

Why oh why hasn't this been discovered before ? Because it's 
much more likely to be a problem when there are:
1) many resources, so the time window grows larger (prod)
2) more concurrent queries going on (prod)