VTK-1313 - Rename storeComment(token, comment) -> addComment(token, comment), + little hack to handle multiple method signatures in RepositoryOperationLogInterceptor.
What we really need to do is reconsider the original addComment(token, resource...) method. Why is the resource a parameter here? Is there a point in retrieving the resource before posting the comment (PostCommentController), sending it to the repository and then retrieving it again by uri to check if it exists? We'll never get this far if it doesn't exist! The comment-object already contains the uri of the resource it belongs to, so the repository only needs a security-token (principal) and a comment-object to check for proper authorization and properly add a comment.

We also need to consider the numberOfComments property. In the case of archiving/expanding, we have a situation where we want to persist comments to the database without affecting this property.