VTK-1044 - Show service name in breadcrumb. Extended breadcumb-component with new parameter "include-service-name". Extended service with servicename-provider. Currently only implemented for tags (vrtx=tags, vrtx=tags&tag=[tag], vrtx=tags&tag=[tag]&view=feed), as a test of concept.
If accepted: must be implemented for other services that require similar extension to the breadcrumb. Also, the different alternate views of our services should use the same provider for setting name/title. E.g. TagsController does not explicitly add the title to the model, title is decided in its corresponding freemarker template (tags.ftl). However, the feed view for tags does exactly this: sets the title on the model.