VTK-1754 Support any number of concurrent read-only index readers (used primarily for searching). This should help improve performance when searching is done by multiple threads, at the cost of higher memory usage and increased dirty-refresh penalty (since potentially multiple instances need refresh when index has become dirty).
Default value is still 1, which is what we have today. Should only be increased
to 2 or perhaps 3 on high-traffic sites where more CPU and memory is available.

Actual gain in concurrent performance remains to be tested. Of course, this
change makes full system testing of query functionality strictly necessary
before putting in production.