Fix VTK-1808.
As it turned out, the #refreshPoolNow() method didn't check if the reader was
actually dirty before submitting to background refresh. And code for background
refresh assumed it would never get a current reader as input. So it didn't check
that a new reader was actually returned by reopen(). It then proceeded to close
it when warmup was finished, in effect closing the current reader and breaking
things.. Fix it by checking that reader is actually dirty in #refreshPoolNow(),
and also add assertive code that checks for this condition in background warm up
task which logs a warning + recovers if it should ever happen.

The race that triggered it:
1. Index was written to, index update-thread called #refreshPoolNow()
2. In the mean time, a reader was refreshed by another request thread through
   getReader(), making it current.
3. #refreshPoolNow() proceeded to dispatch to background warmup without checking
   if readers were in fact now current.
4. Background warmup task closed a current reader, because reopen didn't return
   a new instance, like code exepcted.

I hope there are now no more cases like this, and that the code is stable.