Ok, let's just do it .. Cache generated filters for common "navigation:hidden !exists" ("inverted property exists") queries, which are used as part of many queries building menus, collection lists, etc. Should speed a lot of different things by some unknown, but probably not that large, factor:). Pre-built filters are compact long[] arrays representing doc numbers by bits and don't take up much memory. Cache is keyed on IndexReader instance, but uses weak refs, and so we should be OK memory wise.
Even though it's a specific hack/optimization, I think it's beneficial for our
current situation. Code is unintrusive and can easily be removed if we deem it
unnecessary in the future. We typically do a lot of query work per HTML page we
serve, and the public demands a timely response.