VTK-1896 Do minimal invasion in VortikalServlet, put code into a non-standard error-handler class which logs and writes an HTML page directly to the response.
Add an incident UUID for each lost POST error handled, and display this on error
page while also logging it together with the POST data.

Still a mess, hopefully we can store the POST data in the new session instead,
and replay the POST-to-service once the user has been re-logged in through
whatever authentication mechanism.