The devil is in the details, or something.
Tighten up somewhat and remove some possible corner cases, by
cleaning up management of the lock map and reference counting.

Lock objects are now simpler and don't dispose of themselves by
calling out of inner class, but rather this is handled on the
outside in the main class (methods getLock, returnLock and unlockInternal).