Get rid of custom BinaryContentTypeHandlerCallback since Spring-provided BlobByteArrayTypeHandler can do the job for any database.
This also removes the need to have database-specific versions of "insertPropertyEntry"
statement, so move it into default-sql-maps.xml, and remove it from oracle-functions.xml
and postgresql-functions.xml.

Further, we need to set "wrapAsLob" on DefaultLobHandler for at least PostgreSQL, to make
BlobByteArrayTypeHandler behave as our old BinaryContentTypeHandlerCallback. Oracle
has its own entirely different LobHandler impl and so is not affected.