[VTK-2881] Re-introduce Clob awareness for comment content column in JDBC-handling. Fixes problems with comments > 4000 chars on Oracle.
However, PostgreSQL JDBC4 does not work correctly for Clobs when using
Spring DefaultLobHandler with wrapAsLob set to true. But the option is
required for Blobs to work correctly. So we obviously need to customize
things here - introduce a LobHandlerFactoryBean which creates handler based
on value of our sqlDialect config variable. In that, add a customized
handler for PostgreSQL, which does what we need.