Funny, ConsistencyCheck found an error in itself.
PropertyImpl#equals() does not really cut it for comparing a Property coming from
index and property coming from DAO anymore, since the index has not got the same
amount of info as database. The 'inherited' flag is always false for Property
instances coming from index property sets. So compare only values, and we should
be good.