VTK-3326 - Listing.java, all other altered/created files are a consequence of changes to this particular class.
* Remove list of files (property sets), map of urls (using path as key and full url as value) and boolean array indicating if a given principal has edit privileges on a corresponding property set (indexes of list of files and boolean array must match!!!).

* Replace all the above with a single list containing one entry element (ListingEntry) for each property set. The entry element contains other necessary attributes associated with that property set, such as potential edit privilege and full url.

Also, on a side note, the LinkConstructor does not work well with full urls, so using it to resolve links to particular services on resources from other hosts will be faulty (typically for aggregation). E.g. in image listing where we create links to thumbnail service. Removed in template code for image listing.