Tighten up some racy-looking bits when caching new values.
Previous code got newly cached value back from cache in a later call, to return
it in #get(). Granted, that works most of the time, since likelyhood of expiry
in meantime is slim, but in thoery not, when considering concurrent removal of
older items and cache being maxed out and under high pressure.

The new test case for ContentCache proves this. Fails concurrent testing with
old code (null-values returned by ContentCache#get() under high pressure) and
passes with new code.