* Videoapp callback service. - Allow query of refcount for video objects on host. - Allow reception of update notification messages through POST requests to   an update URI.
Couldn't get Spring REST annotations to work in a reasonable amount of time, so
I did the REST thing manually in a regular Controller (VideoCallbackService).
The controller assumes complete responsibility of its own sub-services, no
formal service XML config at that level of detail, sorry.

* Video event handling
When videoapp callback service receives notifications of updated videos, an
application event is published. The event is handled by VideoUpdateEventHandler,
which is also a system job Task. Events are put on an in-memory queue in request
thread (non-blocking). Events on queue are processed instantly by the job in
task scheduler thread. All resources referencing updated videos are stored,
which causes a refresh of videoref metadata to be done automatically.