Add build of updated Strobe Media Playback from latest regular release. - Patch is slightly modified for robustness, but still does the same   thing of allowing the player to load Vortex thumbnail URLs for video posters. - Built with Adobe Flex 4.5.1 SDK and Flash Builder 4.6.
(Flex 4.6 is default in Flash Builder 4.6, but OSMF/SMP docs mention 4.5.1 and that looks
like the standard for the player, so went with "safest" option. Whatever that is..).

This commit does not update references in Vortikal, so until more testing has been performed, the
previous 1.5.1 version of the player will be used.