[VTK-3627] Make '+' in video paths not break display.
Note that for media references from user input, we cannot know the true
intent of the user when a '+' occurs in the URI, because it is ambiguous.
So we always choose to interpret pluses as literal pluses. This perhaps needs
more consistent handling in general with URIs in Vortex, and one can consider
enforcing this in the URL class.