[VTK-3666] Initial port to Lucene 4.9.0.
It's in a working state, but some things are currently missing due to
extensive differences between Lucene 3 and 4:
- Locale-based sorting for string fields
- Iteration search API with single-field sorting
- Less frequently updated index searcher instances for faster
  anonymous queries are currently omitted, though such a concept might be
  added again, depending on performance testing.

The current state of affairs is only lightly tested in general ! More to
come, and hopefully some new unit tests.

A complete reindexing is necessary, if you're brave enough to test this.