[VTK-3419] First part: make complete Acl available through Index DAO APIs.
Generally, a set of ACL read principals is replaced with a full Acl object,
and is provided by PropertySetHandler through IndexDao. (The actual code
for fetching ACL from database is shared with SqlMapDataAccessor.)

Indexing still only uses read-principals, but now it's derived from the
full ACL instead. Issue not complete with this commit.